Rename from napari-tifffile-reader to napari-tiff (for future addition of writer plugins, etc.)
