+++ docs/index.md
@@ -1,3 +1,3 @@
-# Welcome to napari-tifffile-reader
+# Welcome to napari-tiff
 
 A napari plugin for tiff images.
