+++ mkdocs.yml
@@ -1,10 +1,10 @@
-site_name: napari-tifffile-reader
+site_name: napari-tiff
 site_description: A napari plugin for tiff images.
 site_author: Genevieve Buckley
 
 theme: readthedocs
 
-repo_url: https://github.com/GenevieveBuckley/napari-tifffile-reader
+repo_url: https://github.com/GenevieveBuckley/napari-tiff
 
 pages:
 - Home: index.md
