+++ napari_tiff/napari_tiff_reader.py
